Use TS project references to build docs
