+++ docs/data/base/components/select/UnstyledSelectIntroduction/css/index.tsx
@@ -9,7 +9,7 @@ import { Option, optionClasses } from '@mui/base/Option';
 import { useTheme } from '@mui/system';
 import UnfoldMoreRoundedIcon from '@mui/icons-material/UnfoldMoreRounded';
 import { PopupContext } from '@mui/base/Unstable_Popup';
-import { CssTransition } from '@mui/base';
+import { CssTransition } from '@mui/base/Transitions';
 
 export default function UnstyledSelectIntroduction() {
   return (
