+++ docs/package.json
@@ -12,7 +12,7 @@
     "deploy": "git push -f material-ui-docs master:latest",
     "icons": "rimraf --glob public/static/icons/* && node ./scripts/buildIcons.js",
     "serve": "serve ./export -l 3010",
-    "typescript": "tsc -p tsconfig.json && tsc -p scripts/tsconfig.json",
+    "typescript": "tsc -b tsconfig.json && tsc -p scripts/tsconfig.json",
     "typescript:transpile": "echo 'Use `pnpm docs:typescript:formatted'` instead && exit 1",
     "typescript:transpile:dev": "echo 'Use `pnpm docs:typescript'` instead && exit 1",
     "link-check": "node ./scripts/reportBrokenLinks.js"
