+++ docs/tsconfig.json
@@ -11,7 +11,13 @@
     "skipLibCheck": true,
     "esModuleInterop": true,
     "types": ["react"],
-    "incremental": true
+    "incremental": true,
+    "composite": true
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules"],
+  "references": [
+    {
+      "path": "../packages/mui-base/tsconfig.build.json"
+    }
+  ]
 }
