--- packages/mui-base/src/useAutocomplete/index.d.ts
@@ -1,2 +1,0 @@
-export { useAutocomplete } from './useAutocomplete';
-export * from './useAutocomplete';
