--- packages/mui-base/src/useAutocomplete/index.js
@@ -1,1 +1,0 @@
-export * from './useAutocomplete';
