+++ packages/mui-base/src/useAutocomplete/index.ts
@@ -1,0 +1,2 @@
+export * from './useAutocomplete';
+export * from './useAutocomplete.types';
