+++ tsconfig.json
@@ -14,8 +14,6 @@
     "noErrorTruncation": false,
     "allowJs": true,
     "paths": {
-      "@mui/base": ["./packages/mui-base/src"],
-      "@mui/base/*": ["./packages/mui-base/src/*"],
       "@mui-internal/api-docs-builder": ["./node_modules/@mui/monorepo/packages/api-docs-builder"],
       "@mui-internal/api-docs-builder/*": [
         "./node_modules/@mui/monorepo/packages/api-docs-builder/*"
